server: fix submission queue traversal on request cancellation
vhd_cancel_queued_requests traverses the submission queue using
TAILQ_FOREACH in order to remove and complete the requests matching the
vring.

However, TAILQ_FOREACH is not safe against element removal during the
traversal, which results in crashes or matching elements being skipped.

Open-code the traversal loop to make sure it iterates correctly to the
next element even when the current one is removed.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>