vdev: merge vring.addr_cache with vring.shadow_vq
It looks more consistent if the fields of vring.addr_cache live right in
vring.shadow_vq, as their lifetime is mostly the same.

Also treat .used_gpa_base in the same bundle, which allows to
synchronize its lifetime to the rest of the fields.

No functional changes, just visual consistency improvement.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>