vdev: mark vring stopped after the requests are canceled
When a vring is being stopped in the dataplane, its status is first set
to stopped, and then the requests that remain in flight are canceled.
The cancellation in turn completes every request and decrements the
counter of in flight requests and schedules vring_mark_drain_bh once it
reaches zero if the vring is stopped.

This can lead to the following scenario: if vring_stop_bh is executed
when there are enqueued requests but there are no requests in the
backend, then the last request canceled turns the inflight counter to
zero and, as the vring status is already set to stopped, schedules the
drain BH, but then vring_stop_bh sees the zero inflight counter and
schedules the drain BH again, which breaks the assumptions and fires the
assertion in vring_mark_drained.

Only set the vring status after the requests are canceled, to make sure
the drain BH is only scheduled once.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>