libvhost: add support for multiple request queues
This replaces the signature of the old vhd_register_blockdev() API
and makes it take in an array of queues and a number instead.

The way requests are currently dispatched to queues is just:
rq_idx = vring_idx % num_rqs

Maximum number of queue is defined by VHD_MAX_REQUEST_QUEUES, which is
currently set to 256.

Signed-off-by: Daniil Tatianin <d-tatianin@yandex-team.ru>