blockdev: remove unnecessary request dispatch callback abstraction
The ->dispatch callback would always contain the exact same function
pointer at runtime depending on how it was built. Remove it completely
and instead call the respective dispatch function directly.

Move vblk_handle_request to virtio_blk renaming it to
virtio_blk_handle_request and also mark it as weak so that existing
tests can continue to override it for their own purposes.

Signed-off-by: Daniil Tatianin <d-tatianin@yandex-team.ru>