virtio_blk: don't store bdev pointer
Storing pointer to the object that we don't own is not safe. We are
going to implement "resize" interface, and want to avoid
misunderstanding around the question, should we also modify user-given
bdev object.

So, as a first step, don't store bdev in virtio_blk. Nest step would be
change bdev argument to "const" in public interface.

Signed-off-by: Vladimir Sementsov-Ogievskiy <vsementsov@yandex-team.ru>