server: keep list of all inflight requests in rq
We trusted client to keep track of all requests it dequeued to make sure
all of them are completed, but if anything goes wrong it is too hard to debug
lost or stuck requests so let's keep all inflight requests in list.

This also allows us to easily calculate timestamp of oldest inflight request
that was requested by NBS as useful metrics.

Signed-off-by: Anton Kuchin <antonkuchin@yandex-team.ru>