memmap: make guest memory regions cacheable
We make use of the fact that we're passed an fd for every SET_MEM_TABLE
call, which we can use fstat(2) on to detect whether we alread have this
file descriptor mapped in. Store every mapped memory region in a list,
along with its inode and device ids, as well size and offset. On every
SET_MEM_TABLE call we iterate the region list and attempt to find an
existing mapping of this file descriptor, if one exists we simply
increment its refcount and reuse it without actually mapping anything.

This allows us to save a ton of address space as well as memory and page
fault traffic for VMs that have multiple disks connected to the same
server as we can reuse the same mapping for all disks.

Signed-off-by: Daniil Tatianin <d-tatianin@yandex-team.ru>