Introduce vhost-user-blk-test-server
This is a test vhost-user blk server that uses libvhost combined with
libaio basically acting as a proxy between a virtual disk and an actual
file on disk or a block device.

It has near native bare-metal performance when measured using
fio+io_uring in a linux guest on QEMU vs bare-metal fio on a physical
nvme namespace.

Signed-off-by: Daniil Tatianin <d-tatianin@yandex-team.ru>