tests: allow to override test server path via env
Tests already accept setting test server path via TEST_SERVER_BINARY
environment variable. However it was only effective if the test server
wasn't found at its usual location in the build directory.

Reverse the logic: check the environment first and only use a hard-coded
path as a fallback. This way, a custom libvhost server which doesn't
come bundled can be used in tests. This feels more intuitive as the
environment variable is now an override, not the default.

Signed-off-by: Valentin Sinitsyn <valesini@yandex-team.ru>