tests: build always but make libblkio optional
Disabling tests is a poor practice. Moreover, vhost-user-blk-test-server
is also built together with tests, and it can be of use of its own, even
if the client which comes with libblkio is unavailable.

So refine the logic from commit f616673: build tests unconditionally,
but don't run them is libblkio is disabled.

Signed-off-by: Valentin Sinitsyn <valesini@yandex-team.ru>