catomic.h: allow complilation on arm64
We conducted internal testing of this library
on ARM64 hosts by running a QEMU virtual machine
with a vhost-based boot disk.

The guest booted successfully, and the fio
performance appears good and is comparable
to x86 on a per-queue basis when using a ramfs
backend with fio parameters direct=1 and iodepth=64.

randwrite 4k:
	x86_64 IOPS=49.5k, BW=193MiB/s
	arm64 IOPS=48.0k, BW=188MiB/s

randread 4k:
	x86 IOPS=195k, BW=761MiB/s
	arm64 IOPS=258k, BW=1009MiB/s

read_seq 128k:
	x86: IOPS=36.7k, BW=4588MiB/s
	arm64: IOPS=80.1k, BW=9.78GiB/s

write_seq 128k:
	x86: IOPS=9562, BW=1195MiB/s
	arm64: IOPS=16.7k, BW=2089MiB/s

Signed-off-by: Andrey Smetanin <asm@nebius.com>