vdev: allow calling vdev_handling_finish multiple times
It might be called multiple times in certain error paths in the
following commits. For example in synchronous requests that fail
to send the reply, but also enter the common abort path in conn_read.
The conn_read must still keep its handle_finish because requests may
fail for other reasons besides a broken connection, for example if a
request structure is malformed etc.

Signed-off-by: Daniil Tatianin <d-tatianin@yandex-team.ru>